feat!: Add experimental Create KV store
BREAKING CHANGE!

This removes the `experimentalKVStore` option and replaces it with
`experimentalCreateKVStore`. The old option was broken due to IDB being
used for the database of databases.

This adds a new NamedMemoryStore which can be used to create an in
memory store that shares the data between instances.

TODO: Export the NamedMemoryStore or create an open source repo for it.

Towards #434
Towards #483