chore: Fix tests
Problem 1
----

We were reusing userID and roomID by default, leading to all the tests
using the same IDB instance. This was causing us to call
`onUpdateNeeded` which triggers `location.reload`.

Solution 1
----

- Add a counter to the userID by default. This will cause us to create a
new IDB instance for each test.
- Make the tests fail if we call `onUpdateNeeded`. This will prevent
unintentional calls to `onUpdateNeeded` in the tests.

Problem 2
----

We were not closing Reflect instances in the tests.

Solution 2
----

Keep track of `reflectForTest` instances and close them in `teardown`.

-