chore: Switch to Valita for runtime type validation
Valita has better performance than Superstruct and about the same binary
size.

Valita uses `parse`, but unlike zod it doesn't clone the input object.

Valita `object` does not define whether it allows extra fields or not.
Instead this is a flag that gets passed to `parse`.

The error messages needs some massaging to look better. Will do that as
part of a followup.

Towards #49
Towards #216