fix: Prevent duplicate refs in the chunk
Change the in memory representation of the refs to be a set instead of
an array. This prevents duplicate refs from being added to the chunk.

The IDB chunk format is still using an array. Conceptually we could
change this since IDB uses structured clone which supports Set but our
KV store interface only supports JSON values.