fix: Indexes with prefix should be match the prefix
When we create new index after forking a client we create the indexes by
iterating over the value BTree. When there is a prefix present we start
from the prefix but we forgot to end when the key no longer matches the
prefix, leading the index map to contain items that do not match the
index constraints.