feat(mirror): Find desired semver
We now find the desired semver for the app. This is done by looking for
the dependency of `@rocicorp/reflect` in the `package.json` of the app.

Once found we validate that the semver range is not too restrictive.
Basically, if there is an upper limit it must be a major version (or
minor version for 0.x.x versions).

When we publish the app we now look in Firestore for the newest version
that matches the desired semver range. If we find one we use that
version. If we don't find one we return an error.