feat(mirror): App repo and deployment tracking
When we publish a new app we store the app and the source map in Google
Cloud Storage.

We also create a document in firestore that tracks the deployment of the
app.

If the deployment is successful we update the deployment state to
RUNNING. We also change the state of all the other apps to STOPPED.

If there was an error to publish to Cloudflare we update the status to
FAILED.