chore: Change the app script name
To be friendly words followed by the appID encoded as base32. For
example lavish-medieval-meat-14pc0mi

Previously it was friendly words followed by the appID. The appID is
base62 encoded matching [a-zA-Z0-9] which includes upper case letters.
This is not compatibly with Cloudflares script names.

Now we decode the base62 encoded appID and re-encodes it using base32