feat: Add watch compile
This is not yet hooked up but it has tests to ensure the actual
compiling works.

This is more complicated than I initially anticipated. esbuild has watch
support but it doesn't work when outputting strings. It only works when
outputting files. Instead we use nodejs `fs.watch` to watch the input
files that esbuild reports as dependencies.