feat(reflect-server): Add tail WS endpoint
This is the first step in replacing cloudflare's tail endpoint with our
own.

When a connection is made we override the global `console` with out own
implementation that sends the output to the client over the websocket.

When there are no more connected websockets we restore the original
`console` implementation.