feat(reflect): Bundle things in reflect-cli
reflect-cli has a lot of large and deep dependencies. By letting esbuild
bundle them we can remove them from the dependencies in package.json and
reflect can have a much smaller set of dependencies.

The cli.js is now 8mb (which is quite large) but I decided to not turn
on minification (which halves the size) because at this point we might
want to have reasonable stack traces. Also, the size of the tarball is
not affected as much by minification.