fix(replicache): pendingMutations had wrong IDs
In case there are multiple clients in the client group and more than one
client has pending changes the id was incorrect.

Also, change the order so that the oldest mutation is first in the
returned array.

This is technically a breaking change but this API is flagged as
experimental.