feat(reflect-server): Disconnect beacon plumbing
This adds a new POST endpoint called disconnect. It also referred to as
disconnect beacon to disambiguate it from the disconnect
handler/callback reflect server has.

The disconnect beacon uses the authHandler for authentication. The plan
is that the post body is a pull request body which will be applied.
Currently, the code validates the JSON but it doesn't do anything with
it yet.

This PR also introduces reflect-shared config which allows us to control
features at runtime. The first feature is the `disconnectBeacon`.