fix(replicache: Reject pull/push if closed
There was a bug where if you called pull/push on a closed replicache
instance, it would never resolve or reject. This fixes that so that
the methods reject with an error if the replicache instance is closed.

The push and pull methods returns a promise. If the replicache instance
is closed, then return a rejected promise.