feat(reflect): Add client side disconnect beacon
We now keep track of the last mutationID of the current client. This is
updated when we do a:
- Mutation (as in `rep.mutate.foo()`)
- `refresh`
- `maybeEndPull` when it rebases the mutations after the pull.

This is stored in `#lastMutationID` on the Replicache instance. It is
exposed to Reflect using the `internalOptions.exposeInternalAPI` hook.

In Reflect, we have a disconnect beacon that sends the last mutation ID.
It is called in 2 places:
1. When `Reflect` `close()` is called.
2. When `window` `pagehide` is fired and the page is not persisted in
   the BFCache.