fix(reflect): Add lastMutationIdAtClose to ClientRecord
When the server receives a close beacon it now checks if the client is
still connected.

If the client is still connected, we add the `lastMutationIdAtClose` to
the current client record. This is then checked in the
`processDisconnects` function to see if we can collect the client then.

Another gotcha here is that when we process disconnect we collect the
client even if lastMutationId is larger than lastMutationIdAtClose.
Intuitively this seems correct but this could use some more thought.