chore(reflect): Client tombstone and userID
The ClientRecord now gets a userID. This is the userID that created the
client. When we connect we verify that the userID is the same as the one
that created the client.

When we delete a client we remove the entry under `clientV1/` but we
also write a `clientTombstone/${clientID}` entry.

This also fixes an issue where the closeHandler was not being called
when the client was deleted because it was too old.

Followup to #1418