chore(reflect.net): Use closeHandler
The orchestrator used to listen to pagehide and then call the "unload"
mutator which is a noop on the client but removes the clientID from the
orchestrator on the server.

Instead use the closeHandler which also gets called when the client is
collected which happens on pagehide. It also gets called when the client
is collected if it hasn't been touched in a while.