feat(replicache): In push() and pull() don't reject for network errors.
It's very common for people to call pull() in their poke handler and
not think of the fact that it can reject. In some configurations, this
unhandled rejection causes a pretty dramatic error message.

As of this change, there is no longer a way to know if push()/pull()
completes successfully. If users want this, we can extend the return
value to indicate success without rejecting.