chore(reflect-server): Get rid of client tombstone
Instead have a `deleted` property on the client record.

If a client with pending mutations gets deleted due to inactivity, and
there are other clients in the client group. When these other clients
do a push we need to update the lastMutationID on the server so that
pull can work correctly and mark the pending mutations as applied.