fix(reflect-server): Fast forward and pull logic for deleted clients
We need to update the lmid even if the client is deleted. This is
because when a push happens from a different client in the same client
group we need to mark the mutation as processed in pull or we will get
stuck on mutations done on the deleted client.

For mutations created in the deleted client we ignore the effect of the
mutation but we still update the lastMutationID.