feat(zero-client): Add getQuery and query subscribe callbacks
Basic integration between zero client and zql

This adds a getQuery method which is not the final API but I wanted to
add it as a baby step towards the final API. This method is used to
get an EntityQuery object.

The second part of this PR (should have been a separate PR) is the
making the zql EntityQuery subscribe/unsubscribe call a method on the
Zero client.