fix(replicache): Fix limit for index scans.
We have an optimization in the subscriptions that use scan. This
optimization keeps track of the key that represents the limit. This is
used to determine if a key can affect a subscription that has the scan
with a limit.

However, this optimization is only implemented for non index scans. When
this code was refactored the exit condition for the iterator got
incorrectly merged with the test whether this is an index scan. This
caused the limit to be ignored for index scans.