feat(zql): add set-op for HAVING
this is non-standard SQL but will make everyone's life easier.

Set operations for `HAVING`

E.g.,

```sql
-- issue has all the labels
... HAVING issue.labels SUPERSET ['important', 'urgent']
-- issue has any of the labels
... HAVING issue.labels INTERSECT ['closed', 'cancelled']
-- issue has exactly the labels
... HAVING issue.labels = ['foo']
-- issue has none of the labels
... HAVING issue.labels DISJOINT ['low-pri', 'low', 'unimportant']
```

easiest path forward seems to be to implement these same operators for `WHERE` so we don't have to
create a custom path for `HAVING`.

The extra benefit is it'll make array fields nice to deal with in ZQL.