fix(zql): EntityQuery Return type flattening
Before this change we end up with hard to read types for the `Return`
part of the `EntityQuery` type. This change flattens the return type
to make it easier to read and understand.

For example, given the following query:

```ts
const s = issueQuery
  .join(userQuery, 'x', 'issue.creatorId', 'user.id')
  .select('issue.id', 'issue.title', 'x.id', 'x.name');
```

the `Return` type would be:

```ts
type OldReturnType = ({
    x: {
      name: string;
    };
  } & {
    x: {
      id: string;
    };
  } & {
    issue: {
      title: string;
    };
  } & {
    issue: {
      id: string;
    };
  })[]
>();
```

After this change the `Return` type would be:

```ts
type NewReturnType = {
  x: {
    name: string;
    id: string;
  };
  issue: {
    title: string;
    id: string;
  };
}[];
```