fix(zql): Changes outside the limit window
Previously, changes outside the limit window always added the row.

This was done by collapsing the add and delete operations into a single
`add` operation on the BTree. To keep things simpler and correct, this
optimization has been removed so that we can always ensure min and max
are taking into account and updated as needed.

Towards #1823