chore(replicache): Ensure refs are ordered
and have no duplicates.

This is mostly a way to ensure that we do not do extra work when walking
the DAG. It is not strictly necessary, but it is a good idea.

To keep the code clean I made Refs an opaque type which implies that the
array is sorted and has no duplicates.