feat(zql): move overlay logic into pure functions
So we can re-use this in `TableSource`

I did explore making this its own operator but it needs information that is not available to operators.

The two things that cause headache:
1. `start.basis = before`. This changes how the first element of the stream is determined.
2. `start.row`. This changes whether or not the overlay is included.

Not having access to `start.row` can't be worked around in the case where the `source` is empty. In that case the source emits an empty iterable to its output. An overlay operator wouldn't know what to do with this. Is the current overlay above or below the start of the empty stream as requested by `start.row`?