feat(zql): Add the builder for changing an AST into a pipeline.
This exposed a bug in MemorySource.addOutput(). We were relying
on object identity to tell outputs apart, but the same (join)
operator can be two different outputs.

This made me realize that the interface to Source was wrong. It
should not *be* an Input, but instead should *create* inputs.

This fixes the bug but also makes buildPipeline() nicer by
generalizing all the cases.

It would also allow having each operator call setOutput() on
the prior in its constructor, but this change is already big
enough.

===

It also exposed a bug in MemorySource.push(). When I added
support for indexes, I inadvertently caused one push per-output
and none of the existing tests covered it. Now the builder test
does.