chore(zql): Improve buildPipeline API
Now that Source has its own interface, setOutput can be part of
Input. This allows:

(a) buildPipeline to treat the whole chain generically, as it was
in ivm1.

(b) Each operator can call setOutput() on the prior in its
constructor. Not sure what this means for OR, but we can cross
that bridge when we get there.

(c) The interface to buildPipeline() can be prettier - we can
just return the last operator and let user build their sink,
rather than requiring a factory.