feat(zero): Add a View class.
View maintains a materialized view of the output of a pipeline.

The view is maintained in a b-tree. Users can listen for changes
and they receive an iterable. The iterable is invalidated as soon
as the listener exits, so users will likely have to copy to some
other data structure. But this would be the case no matter what
in React at least, since React needs to be able to re-render at
any time.

This class leaves this final copying step to the framework-
specific bindings layer above.

This class also changes the output type from Node to "Entry",
which smooshes together the Row and its relationships for
ergonomics.

I hadn't thought carefully about the output interface before but
of course this is what people would want. I was mainly focused
on making the internals clean and simple and minimizing copies,
but the presence of relationships makes copies a lot harder to
avoid anyway.