chore: Change IVM ChangeType to SMI
For performance reasons, we are changing the IVM ChangeType from a
string to a SMI. SMIs are stack allocated and stored inline in objects
so this should be slightly better.

But, another big reason is to have TypeScript detect missing cases as we
add ChangeType.Edit in an upcoming change.