feat(zql): Add edit change type
When a change diff is made by Replicache, we no longer create a
RemoveChange followed by an AddChange. Instead we create an EditChange
that has the old and new rows. This is more efficient and it allows us
to do less work in Join.

With the old RemoveChange/AddChange model, we would always tear down the
row with the relation and then recreate the row and all its subrows.
With the new EditChange model, we can just update the row and its
subrows in place in the case where the join key hasn't changed.

This is not yet enabled by default because not all operators are done.

- [x] Join
- [x] ArrayView
- [x] MemorySource
- [x] Catch
- [x] Snitch
- [ ] Filter
- [ ] Skip
- [ ] Take