chore: Enable verbatimModuleSyntax in tsconfig
This is the setting that tells us to use type imports for types.

We used to have this using `importsNotUsedAsValues` but the option was
removed.

This is the new way to do it and it is also a bit more strict.

This lead to a few minor manual changes which I'll comment on in the PR