chore(zero): Change the dispatcher listens on to default.
Fastify sets this to localhost and IIUC this will make it listen on
both ipv4 and ipv6 loopback addresses.

I don't know how this makes it accessible from other computers on
AWS as localhost is supposed to be restricted to only local machine
but apparently it works so 🤷‍♂️. Maybe dax will see this commit
go by and explain.