chore(zbugs): One more crack at shortIDs.
Matt brought of the question of what happens if a user gets ahold
of a long ID? Do they have to know about '/pending' ?

I don't really buy this argument. The shortID is nicer - why would
a user have a long ID in the first place.

But say it was like development or something, it's easy enough to
remember the URL convention.

That said, I don't like how much the /pending/<longID> URL flashes,
and having the two routes is slightly annoying. So using a single
char prefix to distinguish the two cases instead.

Yes I can just use isNumeric() and yes the chances of it getting
the wrong thing are astronomical. But I like not having to think
about that math.