feat(zero)!: Where expression factory
BREAKING CHANGE!

The `Query` `where` method now accepts a lambda function that returns a
`Condition`. The function gets passed one argument which is an
`ExpressionBuilder`. The `ExpressionBuilder` carries the `TableSchema`
type so that `cmp` calls can be type checked.

For example:

```ts
const query = z.query.user.where(({and, cmp, or}) =>
  and(
    cmp('name', '=', 'Alice'),
    or(cmp('age', '>', 18), cmp('isStudent', '=', true)),
  ));
```

This removes the free `and`, `cmp`, `not` and `or` functions fro the
`@rocicoirp/zero` package. This removal is a breaking change.

Towards https://bugs.rocicorp.dev/issue/3104