fix(zero): The order of the primary key matters
This was an oversight by me. If the pk is `['b', 'a']` the source is
first sorted by `'b'` and then by `'a'`. This is not the same as sorting
by `'a'` and then by `'b'`. This commit fixes the issue by sorting the
primary key in the correct order.