feat(zbugs): Top and bottom toast position
Determine where the emoji is relative to the current viewport. If the
emoji is above, show the toast at the top. If the emoji is below, show
the toast at the bottom. If the emoji is in the viewport do not show
the toast.

There is one minor issue and that is that if the comment has 1px below
the bottom edge of the viewport, the toast will be shown at the bottom
of the viewport. This is a minor issue and will NOT be addressed.