feat(zbugs): Show a toast when a new comment is added
This is done by keeping track of the comment IDs in a Ref. When the
comment IDs change we show a toast.

For the emojis we had an issue where the issue-page was reused between
issues. This is not causing issues here which is not clear to me yet
why that is not the case.