fix(zbugs): Issue count badge tweaks
This is to prevent if from changing size while we wait for results.

If we have results or the result type is complete we show the badge,
otherwise we hide it. This implies that it is not shown the first frame
when refreshing the page.

Fixes https://bugs.rocicorp.dev/issue/3249