fix(zero-schema): Fix an issue with TS2589
TS2589: Type instantiation is excessively deep and possibly infinite

The hint to solve this was here:

https://github.com/microsoft/TypeScript/issues/30188#issuecomment-478938437

The solution is to compute the types that cause problems step by step.
By using `T extends infer U`, we can force the computation of `T`
instead of deferring it. This way the instantiation does not get deeper.

 instantiation is excessively deep and possibly infinite