fix(zql,z2s): place `hidden` flag on the correct relationship
query-impl put the `hidden` flag at the final table in a junction edge, which was incorrect. ZQL did the right thing because `join` moved the hidden flag up a level incorrectly. So two errors made a right in this case.

The z2s compiler, however, couldn't work properly with the `hidden` flag in the wrong level from `query-impl`