chore(zero): Remove syncing of connected clients
We used to sync the clientIDs of the connected clients. The idea was
that we would use this information to determine which clients we cannot
delete. However, this information is not used anywhere in the codebase
and is not necessary. This commit removes the syncing of the clientIDs.

This means that we can remove the clientPatches that we send.

Since we are not sending client patches we can also remove this state
from the CRV.