feat(zero-cache): Implement server side of deleteClients
When the server receives a DeleteClients message, it will delete the
client (and client group) from the CRV database. This will send
`'desiredQueriesPatches'` and `'gotQueriesPatch'` patches to connected
clients that are affected by the deletions.

The server also confirms the deletion by sending a `'deleteClients'`
back to the connected clients that requested the deletion.

Towards https://bugs.rocicorp.dev/issue/3476