feat(zero-cache): Evict inactive queries
We now evict inactive queries from the cache. This is done by keeping
track of the time a query was inactivated. When the number of rows in
the RowRecordCache is larger than the desired rows then we evict the
oldest inactive query. After that the run loop starts over and we might
evict another one.